chore: Script improvements
This performs a few improvements:
- Moves the Script class and interface to the Core directory
- Performs consistency updates
