+++ src/Aggregation/Metrics/TopHitsAggregation.php
@@ -5,10 +5,10 @@ declare(strict_types = 1);
 namespace Hypefactors\ElasticBuilder\Aggregation\Metrics;
 
 use Hypefactors\ElasticBuilder\Aggregation\Aggregation;
+use Hypefactors\ElasticBuilder\Core\ScriptInterface;
 use Hypefactors\ElasticBuilder\Core\SortInterface;
 use Hypefactors\ElasticBuilder\Core\Util;
 use Hypefactors\ElasticBuilder\Highlight\HighlightInterface;
-use Hypefactors\ElasticBuilder\Script\ScriptInterface;
 use InvalidArgumentException;
 
 /**
@@ -158,8 +158,8 @@ class TopHitsAggregation extends Aggregation
     /**
      * Allows to return a script evaluation (based on different fields) for each hit.
      *
-     * @param string                                             $fieldName
-     * @param \Hypefactors\ElasticBuilder\Script\ScriptInterface $script
+     * @param string                                           $fieldName
+     * @param \Hypefactors\ElasticBuilder\Core\ScriptInterface $script
      *
      * @return $this
      */
