+++ tests/Core/InnerHitsTest.php
@@ -9,8 +9,8 @@ use Hypefactors\ElasticBuilder\Core\Sort;
 use Hypefactors\ElasticBuilder\Core\SortInterface;
 use Hypefactors\ElasticBuilder\Highlight\Highlight;
 use Hypefactors\ElasticBuilder\Highlight\HighlightInterface;
-use Hypefactors\ElasticBuilder\Script\Script;
-use Hypefactors\ElasticBuilder\Script\ScriptInterface;
+use Hypefactors\ElasticBuilder\Core\Script;
+use Hypefactors\ElasticBuilder\Core\ScriptInterface;
 use PHPUnit\Framework\TestCase;
 use stdClass;
