+++ tests/Core/ScriptTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types = 1);
 
-namespace Hypefactors\ElasticBuilder\Tests\Script;
+namespace Hypefactors\ElasticBuilder\Tests\Core;
 
-use Hypefactors\ElasticBuilder\Script\Script;
+use Hypefactors\ElasticBuilder\Core\Script;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
